converting par in plotspict.priors to matrix in case that numeric
plotspict.priors() was not working when setting a prior to Biomass or Fishing mortality or both, because `par` was returned as numeric. Changes make sure that `nrow`  and subsetting work on `par` by converting `par` to matrix under the assumption that `ncol` is always 5 (ll, eat, ul, sd, cv). You might find a better solution but this worked for me, with every prior combination.