updating check for large uncertainties in `plotspict.biomass`
Fixing the calculation of `fininds` in `plotspict.biomass` to 5 may evoke an error by returning an empty variable which is later used as index. This occurs when no cv is smaller than 5, as it was in my data. Implemented changes use 5 as the cv threshold only if any cv value smaller than 5 exists, otherwise the minimum cv is used to have at least one value limiting the scaling of the plot.